decomposing try_execute_fusion_group
